pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "repl_python3_C"
version = "0.1.0"
[tool.poetry.dependencies]
Flask = "^2.0.2"
python = "^3.8"
transformers = {extras = ["tf-cpu"], version = "^4.12.3"}
PyJWT = "^2.3.0"
pickle4 = "^0.0.1"
sklearn = "^0.0"
[tool.poetry.dev-dependencies]
